backend/requirements.txt
fastapi==0.115.0
mangum==0.19.0
pydantic==2.9.2
httpx==0.27.2
uvicorn==0.30.6
boto3==1.35.37
pyjwt==2.9.0
pytest==8.3.3
respx==0.22.0
websockets==12.0
protobuf==4.25.0
grpcio-tools==1.60.0
